+++ frontend/src/views/AcquisitionView.tsx
