+++ frontend/src/views/TemplatesView.tsx
